+++ src/CfdiUtils/Elements/Dividendos10/DividOUtil.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Dividendos10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class DividOUtil extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'dividendos:DividOUtil';
+    }
+}