+++ src/CfdiUtils/Elements/Dividendos10/Remanente.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Dividendos10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Remanente extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'dividendos:Remanente';
+    }
+}