+++ tests/CfdiUtilsTests/Elements/Dividendos10/RemanenteTest.php
@@ -1,0 +1,22 @@
+<?php
+namespace CfdiUtilsTests\Elements\Dividendos10;
+
+use CfdiUtils\Elements\Dividendos10\Remanente;
+use PHPUnit\Framework\TestCase;
+
+class RemanenteTest extends TestCase
+{
+    /** @var Remanente */
+    public $element;
+
+    public function setUp()
+    {
+        parent::setUp();
+        $this->element = new Remanente();
+    }
+
+    public function testGetElementName()
+    {
+        $this->assertSame('dividendos:Remanente', $this->element->getElementName());
+    }
+}